server: Simplify multi-step operation handling
Previously, multi-step operations like registering and logging in were handled
by having the client store a "session" id cookie after the initial request,
only to send it back in the finalizing request seconds later in order to be
linked up with the challenge stored for the initial request.

This can be simplified in two ways:
- We don't need any "session" id cookie at all, because the challenge serves that
  very purpose already, and the challenge is sent along with the response
- We don't need a whole model for updating "session" storage, because
  it's really not a session at all, but rather just a link between two
  requests. In the initial requests we just need to store the challenge,
  and delete it again in the finalizing request, no updates needed.

This commit also renames this server state from the previous session
state to pending operations, because it really isn't very much like a
session.

This commit also allows already-logged in users to re-authenticate.